+++ apps/explorer/lib/explorer/chain/import/stage/block_referencing.ex
@@ -13,6 +13,7 @@ defmodule Explorer.Chain.Import.Stage.BlockReferencing do
     Runner.Logs,
     Runner.Tokens,
     Runner.TokenTransfers,
+    Runner.TokenInstances,
     Runner.Address.TokenBalances,
     Runner.TransactionActions,
     Runner.Withdrawals